Remove duplicate sql utils & validator code
Fixes #2628
