+++ azure-cli.pyproj
@@ -341,7 +341,6 @@
     <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\custom.py" />
     <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\help.py" />
     <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\params.py" />
-    <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\validators.py" />
     <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\_util.py" />
     <Compile Include="command_modules\azure-cli-sql\azure\cli\command_modules\sql\__init__.py" />
     <Compile Include="command_modules\azure-cli-sql\setup.py" />
